food_tracker/pyproject.toml
[tool.poetry]
name = "food_tracker"
version = "0.1.0"
description = ""
authors = ["Roman Kosarev <rmksrv@outlook.com>"]

[tool.poetry.dependencies]
python = "^3.10"
Flask = "^2.2.2"
python-dateutil = "^2.8.2"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
